refactor plot into submodule
