+++ python/scprep/plot/__init__.py
@@ -1,0 +1,4 @@
+from .scatter import scatter, scatter2d, scatter3d, rotate_scatter3d
+from .histogram import histogram, plot_library_size, plot_gene_set_expression
+from .scree import scree_plot
+from . import tools
