+++ python/scprep/plot/scatter.py
@@ -1,6 +1,4 @@
 import numpy as np
-from decorator import decorator
-import platform
 import numbers
 import pandas as pd
 import warnings
@@ -8,360 +6,12 @@ try:
     import matplotlib.pyplot as plt
     import matplotlib as mpl
     from matplotlib import animation
-    from mpl_toolkits.mplot3d import Axes3D
 except ImportError:
     pass
 
-from . import measure, utils
-
-
-@decorator
-def _with_matplotlib(fun, *args, **kwargs):
-    try:
-        plt
-    except NameError:
-        raise ImportError(
-            "matplotlib not found. "
-            "Please install it with e.g. `pip install --user matplotlib`")
-    return fun(*args, **kwargs)
-
-
-def _mpl_is_gui_backend():
-    backend = mpl.get_backend()
-    if backend in ['module://ipykernel.pylab.backend_inline', 'agg']:
-        return False
-    else:
-        return True
-
-
-def _get_figure(ax=None, figsize=None, subplot_kw=None):
-    if subplot_kw is None:
-        subplot_kw = {}
-    if ax is None:
-        fig, ax = plt.subplots(figsize=figsize, subplot_kw=subplot_kw)
-        show_fig = True
-    else:
-        try:
-            fig = ax.get_figure()
-        except AttributeError as e:
-            if not isinstance(ax, mpl.axes.Axes):
-                raise TypeError("Expected ax as a matplotlib.axes.Axes. "
-                                "Got {}".format(type(ax)))
-            else:
-                raise e
-        if 'projection' in subplot_kw:
-            if subplot_kw['projection'] == '3d' and not isinstance(ax, Axes3D):
-                raise TypeError("Expected ax with projection='3d'. "
-                                "Got 2D axis instead.")
-        show_fig = False
-    return fig, ax, show_fig
-
-
-def _is_color_array(c):
-    return c is not None and np.all([mpl.colors.is_color_like(val) for val in c])
-
-
-def _in_ipynb():
-    """Check if we are running in a Jupyter Notebook
-
-    Credit to https://stackoverflow.com/a/24937408/3996580
-    """
-    __VALID_NOTEBOOKS = ["<class 'google.colab._shell.Shell'>",
-                         "<class 'ipykernel.zmqshell.ZMQInteractiveShell'>"]
-    try:
-        return str(type(get_ipython())) in __VALID_NOTEBOOKS
-    except NameError:
-        return False
-
-
-@_with_matplotlib
-def show(fig):
-    """Show a matplotlib Figure correctly, regardless of platform
-
-    If running a Jupyter notebook, we avoid running `fig.show`. If running
-    in Windows, it is necessary to run `plt.show` rather than `fig.show`.
-
-    Parameters
-    ----------
-    fig : matplotlib.Figure
-        Figure to show
-    """
-    if _mpl_is_gui_backend():
-        plt.tight_layout()
-        if platform.system() == "Windows":
-            plt.show(block=True)
-        else:
-            fig.show()
-
-
-@_with_matplotlib
-def histogram(data,
-              bins=100, log=True,
-              cutoff=None, percentile=None,
-              ax=None, figsize=None,
-              xlabel=None,
-              ylabel='Number of cells',
-              **kwargs):
-    """Plot a histogram.
-
-    Parameters
-    ----------
-    data : array-like, shape=[n_samples]
-        Input data
-    bins : int, optional (default: 100)
-        Number of bins to draw in the histogram
-    log : bool, or {'x', 'y'}, optional (default: True)
-        If True, plot both axes on a log scale. If 'x' or 'y',
-        only plot the given axis on a log scale. If False,
-        plot both axes on a linear scale.
-    cutoff : float or `None`, optional (default: `None`)
-        Absolute cutoff at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    percentile : float or `None`, optional (default: `None`)
-        Percentile between 0 and 100 at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    ax : `matplotlib.Axes` or None, optional (default: None)
-        Axis to plot on. If None, a new axis will be created.
-    figsize : tuple or None, optional (default: None)
-        If not None, sets the figure size (width, height)
-    [x,y]label : str, optional
-        Labels to display on the x and y axis.
-    **kwargs : additional arguments for `matplotlib.pyplot.hist`
-
-    Returns
-    -------
-    ax : `matplotlib.Axes`
-        axis on which plot was drawn
-    """
-    fig, ax, show_fig = _get_figure(ax, figsize)
-    if log == 'x' or log is True:
-        bins = np.logspace(np.log10(max(np.min(data), 1)),
-                           np.log10(np.max(data)),
-                           bins)
-    ax.hist(data, bins=bins, **kwargs)
-
-    if log == 'x' or log is True:
-        ax.set_xscale('log')
-    if log == 'y' or log is True:
-        ax.set_yscale('log')
-
-    if xlabel is not None:
-        ax.set_xlabel(xlabel)
-    if ylabel is not None:
-        ax.set_ylabel(ylabel)
-
-    cutoff = measure._get_percentile_cutoff(
-        data, cutoff, percentile, required=False)
-    if cutoff is not None:
-        ax.axvline(cutoff, color='red')
-    if show_fig:
-        show(fig)
-    return ax
-
-
-@_with_matplotlib
-def plot_library_size(data,
-                      bins=100, log=True,
-                      cutoff=None, percentile=None,
-                      ax=None, figsize=None,
-                      xlabel='Library size',
-                      **kwargs):
-    """Plot the library size histogram.
-
-    Parameters
-    ----------
-    data : array-like, shape=[n_samples, n_features]
-        Input data
-    bins : int, optional (default: 100)
-        Number of bins to draw in the histogram
-    log : bool, or {'x', 'y'}, optional (default: True)
-        If True, plot both axes on a log scale. If 'x' or 'y',
-        only plot the given axis on a log scale. If False,
-        plot both axes on a linear scale.
-    cutoff : float or `None`, optional (default: `None`)
-        Absolute cutoff at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    percentile : float or `None`, optional (default: `None`)
-        Percentile between 0 and 100 at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    ax : `matplotlib.Axes` or None, optional (default: None)
-        Axis to plot on. If None, a new axis will be created.
-    figsize : tuple or None, optional (default: None)
-        If not None, sets the figure size (width, height)
-    [x,y]label : str, optional
-        Labels to display on the x and y axis.
-    **kwargs : additional arguments for `matplotlib.pyplot.hist`
-
-    Returns
-    -------
-    ax : `matplotlib.Axes`
-        axis on which plot was drawn
-    """
-    return histogram(measure.library_size(data),
-                     cutoff=cutoff, percentile=percentile,
-                     bins=bins, log=log, ax=ax, figsize=figsize,
-                     xlabel=xlabel, **kwargs)
-
-
-@_with_matplotlib
-def plot_gene_set_expression(data, genes,
-                             bins=100, log=False,
-                             cutoff=None, percentile=None,
-                             library_size_normalize=True,
-                             ax=None, figsize=None,
-                             xlabel='Gene expression',
-                             **kwargs):
-    """Plot the histogram of the expression of a gene set.
-
-    Parameters
-    ----------
-    data : array-like, shape=[n_samples, n_features]
-        Input data
-    genes : list-like, dtype=`str` or `int`
-        Integer column indices or string gene names included in gene set
-    bins : int, optional (default: 100)
-        Number of bins to draw in the histogram
-    log : bool, or {'x', 'y'}, optional (default: True)
-        If True, plot both axes on a log scale. If 'x' or 'y',
-        only plot the given axis on a log scale. If False,
-        plot both axes on a linear scale.
-    cutoff : float or `None`, optional (default: `None`)
-        Absolute cutoff at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    percentile : float or `None`, optional (default: `None`)
-        Percentile between 0 and 100 at which to draw a vertical line.
-        Only one of `cutoff` and `percentile` may be given.
-    library_size_normalize : bool, optional (default: True)
-        Divide gene set expression by library size
-    ax : `matplotlib.Axes` or None, optional (default: None)
-        Axis to plot on. If None, a new axis will be created.
-    figsize : tuple or None, optional (default: None)
-        If not None, sets the figure size (width, height)
-    [x,y]label : str, optional
-        Labels to display on the x and y axis.
-    **kwargs : additional arguments for `matplotlib.pyplot.hist`
-
-    Returns
-    -------
-    ax : `matplotlib.Axes`
-        axis on which plot was drawn
-    """
-    return histogram(measure.gene_set_expression(
-        data, genes, library_size_normalize=library_size_normalize),
-        cutoff=cutoff, percentile=percentile,
-        bins=bins, log=log, ax=ax, figsize=figsize,
-        xlabel=xlabel, **kwargs)
-
-
-@_with_matplotlib
-def scree_plot(singular_values, cumulative=False, ax=None, figsize=None,
-               xlabel='Principal Component', ylabel='Explained Variance (%)',
-               **kwargs):
-    """Plot the explained variance of each principal component
-
-    Parameters
-    ----------
-    singular_values : list-like, shape=[n_components]
-        Singular values returned by `scprep.reduce.pca(data, return_singular_values=True)`
-    cumulative : bool, optional (default=False)
-        If True, plot the cumulative explained variance
-    ax : `matplotlib.Axes` or None, optional (default: None)
-        Axis to plot on. If None, a new axis will be created.
-    figsize : tuple or None, optional (default: None)
-        If not None, sets the figure size (width, height)
-    {x,y}label : str, optional
-        Labels to display on the x and y axis.
-    **kwargs : additional arguments for `matplotlib.pyplot.plot`
-
-    Returns
-    -------
-    ax : `matplotlib.Axes`
-        axis on which plot was drawn
-
-    Examples
-    --------
-    >>> import scprep
-    >>> import numpy as np
-    >>> data = np.random.normal(0, 1, [200, 1000])
-    >>> pca_data, singular_values = scprep.reduce.pca(data, n_components=100, return_singular_values=True)
-    >>> scprep.plot.scree_plot(singular_values)
-    >>> scprep.plot.scree_plot(singular_values, cumulative=True)
-    """
-    explained_variance = singular_values ** 2
-    explained_variance = explained_variance / explained_variance.sum()
-    if cumulative:
-        explained_variance = np.cumsum(explained_variance)
-    fig, ax, show_fig = _get_figure(ax, figsize)
-    ax.plot(np.arange(len(explained_variance)), explained_variance, **kwargs)
-    ax.set_xlabel(xlabel)
-    ax.set_ylabel(ylabel)
-    if show_fig:
-        show(fig)
-    return ax
-
-
-@_with_matplotlib
-def create_colormap(colors, name="scprep_custom_cmap"):
-    """Create a custom colormap from a list of colors
-
-    Parameters
-    ----------
-    colors : list-like
-        List of `matplotlib` colors. Includes RGB, RGBA,
-        string color names and more.
-        See <https://matplotlib.org/api/colors_api.html>
-
-    Returns
-    -------
-    cmap : `matplotlib.colors.LinearSegmentedColormap`
-        Custom colormap
-    """
-    if len(colors) == 1:
-        colors = np.repeat(colors, 2)
-    vals = np.linspace(0, 1, len(colors))
-    cdict = dict(red=[], green=[], blue=[], alpha=[])
-    for val, color in zip(vals, colors):
-        r, g, b, a = mpl.colors.to_rgba(color)
-        cdict['red'].append((val, r, r))
-        cdict['green'].append((val, g, g))
-        cdict['blue'].append((val, b, b))
-        cdict['alpha'].append((val, a, a))
-    cmap = mpl.colors.LinearSegmentedColormap(name, cdict)
-    return cmap
-
-
-def _create_normalize(vmin, vmax, scale="linear"):
-    """Create a colormap normalizer
-
-    Parameters
-    ----------
-    scale : {'linear', 'log', 'symlog', 'sqrt'} or `matplotlib.colors.Normalize`, optional (default: 'linear')
-        Colormap normalization scale. For advanced use, see
-        <https://matplotlib.org/users/colormapnorms.html>
-
-    Returns
-    -------
-    norm : `matplotlib.colors.Normalize`
-    """
-    if scale == 'linear':
-        norm = mpl.colors.Normalize(vmin=vmin, vmax=vmax)
-    elif scale == 'log':
-        if vmin <= 0:
-            raise ValueError(
-                "`vmin` must be positive for `cmap_scale='log'`. Got {}".format(vmin))
-        norm = mpl.colors.LogNorm(vmin=vmin, vmax=vmin)
-    elif scale == 'symlog':
-        norm = mpl.colors.SymLogNorm(linthresh=0.03, linscale=0.03,
-                                     vmin=vmin, vmax=vmax)
-    elif scale == 'sqrt':
-        norm = mpl.colors.PowerNorm(gamma=1. / 2.)
-    elif isinstance(scale, mpl.colors.Normalize):
-        norm = scale
-    else:
-        raise ValueError("Expected norm in ['linear', 'log', 'symlog',"
-                         "'sqrt'] or a matplotlib.colors.Normalize object."
-                         " Got {}".format(scale))
-    return norm
+from .. import utils
+from .utils import _get_figure, _with_matplotlib, _is_color_array, show, _in_ipynb
+from .tools import create_colormap, create_normalize, label_axis, generate_colorbar, generate_legend
 
 
 class _ScatterParams(object):
@@ -555,7 +205,7 @@ class _ScatterParams(object):
     @property
     def norm(self):
         if self.cmap_scale is not None and self.cmap_scale != 'linear':
-            return _create_normalize(self.vmin, self.vmax, self.cmap_scale)
+            return create_normalize(self.vmin, self.vmax, self.cmap_scale)
         else:
             return None
 
@@ -682,160 +332,6 @@ class _ScatterParams(object):
                     "Cannot use non-linear `cmap_scale` with discrete data.",
                     UserWarning)
                 self._cmap_scale = 'linear'
-
-
-@_with_matplotlib
-def generate_legend(cmap, ax, title=None, marker='o', markersize=10,
-                    loc='best', bbox_to_anchor=None,
-                    fontsize=14, title_fontsize=14,
-                    max_rows=10, ncol=None, **kwargs):
-    """Generate a legend on an axis.
-
-    Parameters
-    ----------
-    cmap : dict
-        Dictionary of label-color pairs.
-    ax : `matplotlib.axes.Axes`
-        Axis on which to draw the legend
-    title : str, optional (default: None)
-        Title to display alongside colorbar
-    marker : str, optional (default: 'o')
-        `matplotlib` marker to use for legend points
-    markersize : float, optional (default: 10)
-        Size of legend points
-    loc : int or string or pair of floats, default: 'best'
-        Matplotlib legend location.
-        See <https://matplotlib.org/api/_as_gen/matplotlib.pyplot.legend.html>
-        for details.
-    bbox_to_anchor : `BboxBase`, 2-tuple, or 4-tuple
-        Box that is used to position the legend in conjunction with loc.
-        See <https://matplotlib.org/api/_as_gen/matplotlib.pyplot.legend.html>
-        for details.
-    fontsize : int, optional (default: 14)
-        Font size for legend labels
-    title_fontsize : int, optional (default: 14)
-        Font size for legend title
-    max_rows : int, optional (default: 10)
-        Maximum number of labels in a column before overflowing to
-        multi-column legend
-    ncol : int, optional (default: None)
-        Number of legend columns. Overrides `max_rows`.
-    kwargs : additional arguments for `plt.legend`
-
-    Returns
-    -------
-    legend : `matplotlib.legend.Legend`
-    """
-    handles = [mpl.lines.Line2D([], [], marker=marker, color=color,
-                                linewidth=0, label=label,
-                                markersize=markersize)
-               for label, color in cmap.items()]
-    if ncol is None:
-        ncol = max(1, len(cmap) // max_rows)
-    legend = ax.legend(handles=handles, title=title,
-                       loc=loc, bbox_to_anchor=bbox_to_anchor,
-                       fontsize=fontsize, ncol=ncol, **kwargs)
-    plt.setp(legend.get_title(), fontsize=title_fontsize)
-    return legend
-
-
-@_with_matplotlib
-def generate_colorbar(cmap, vmin=None, vmax=None, scale='linear', ax=None,
-                      title=None, title_fontsize=12, title_rotation=270,
-                      **kwargs):
-    """Generate a colorbar on an axis.
-
-    Parameters
-    ----------
-    cmap : `matplotlib` colormap or str
-        Colormap with which to draw colorbar
-    scale : {'linear', 'log', 'symlog', 'sqrt'} or `matplotlib.colors.Normalize`, optional (default: 'linear')
-        Colormap normalization scale. For advanced use, see
-        <https://matplotlib.org/users/colormapnorms.html>
-    vmin, vmax : float, optional (default: None)
-        Range of values to display on colorbar
-    ax : `matplotlib.axes.Axes`, list or None, optional (default: None)
-        Axis or list of axes from which to steal space for colorbar
-        If `None`, uses the current axis
-    title : str, optional (default: None)
-        Title to display alongside colorbar
-    title_fontsize : int, optional (default: 14)
-        Font size for colorbar title
-    title_rotation : int, optional (default: 270)
-        Angle of rotation of the colorbar title
-    kwargs : additional arguments for `plt.colorbar`
-
-    Returns
-    -------
-    colorbar : `matplotlib.colorbar.Colorbar`
-    """
-    try:
-        plot_axis = ax[0]
-    except TypeError:
-        # not a list
-        plot_axis = ax
-    if vmax is None and vmin is None:
-        vmax = 1
-        vmin = 0
-        remove_ticks = True
-        norm = None
-    elif vmax is None or vmin is None:
-        raise ValueError("Either both or neither of `vmax` and `vmin` should "
-                         "be set. Got `vmax={}, vmin={}`".format(vmax, vmin))
-    else:
-        remove_ticks = False
-        norm = _create_normalize(vmin, vmax, scale=scale)
-    fig, plot_axis, _ = _get_figure(plot_axis)
-    if ax is None:
-        ax = plot_axis
-    xmin, xmax = plot_axis.get_xlim()
-    ymin, ymax = plot_axis.get_ylim()
-    im = plot_axis.imshow(np.linspace(vmin, vmax, 10).reshape(-1, 1),
-                          vmin=vmin, vmax=vmax, cmap=cmap, norm=norm,
-                          aspect='auto', origin='lower',
-                          extent=[xmin, xmax, ymin, ymax])
-    im.remove()
-    colorbar = fig.colorbar(im, ax=ax, **kwargs)
-    if title is not None:
-        colorbar.set_label(title, rotation=title_rotation,
-                           fontsize=title_fontsize)
-    if remove_ticks:
-        colorbar.set_ticks([])
-    return colorbar
-
-
-def _label_axis(axis, ticks=True, ticklabels=True, label=None):
-    """Set axis ticks and labels
-
-    Parameters
-    ----------
-    axis : matplotlib.axis.{X,Y}Axis, mpl_toolkits.mplot3d.axis3d.{X,Y,Z}Axis
-        Axis on which to draw labels and ticks
-    ticks : True, False, or list-like (default: True)
-        If True, keeps default axis ticks.
-        If False, removes axis ticks.
-        If a list, sets custom axis ticks
-    ticklabels : True, False, or list-like (default: True)
-        If True, keeps default axis tick labels.
-        If False, removes axis tick labels.
-        If a list, sets custom axis tick labels
-    label : str or None (default : None)
-        Axis labels. If None, no label is set.
-    """
-    if not ticks:
-        axis.set_ticks([])
-    elif ticks is True:
-        pass
-    else:
-        axis.set_ticks(ticks)
-    if not ticklabels:
-        axis.set_ticklabels([])
-    elif ticklabels is True:
-        pass
-    else:
-        axis.set_ticklabels(ticklabels)
-    if label is not None:
-        axis.set_label_text(label)
 
 
 @_with_matplotlib
@@ -993,10 +489,10 @@ def scatter(x, y, z=None,
             zlabel = label_prefix + "3"
 
     # label axes
-    _label_axis(ax.xaxis, xticks, xticklabels, xlabel)
-    _label_axis(ax.yaxis, yticks, yticklabels, ylabel)
+    label_axis(ax.xaxis, xticks, xticklabels, xlabel)
+    label_axis(ax.yaxis, yticks, yticklabels, ylabel)
     if z is not None:
-        _label_axis(ax.zaxis, zticks, zticklabels, zlabel)
+        label_axis(ax.zaxis, zticks, zticklabels, zlabel)
 
     if title is not None:
         ax.set_title(title)
